+++ src/pages/namespace/Explorer/Endpoint/EndpointEditor/Form/plugins/components/Modal.tsx
@@ -8,7 +8,6 @@ import { FC, PropsWithChildren } from "react";
 
 import Button from "~/design/Button";
 import { Card } from "~/design/Card";
-import { twMergeClsx } from "~/util/helpers";
 import { useTranslation } from "react-i18next";
 
 type ModalWrapperProps = PropsWithChildren & {
@@ -42,34 +41,6 @@ export const PluginWrapper: FC<PropsWithChildren> = ({ children }) => (
   <Card className="flex flex-col gap-5 p-5" noShadow>
     {children}
   </Card>
-);
-
-type FieldsetProps = PropsWithChildren & {
-  label: string;
-  htmlFor?: string;
-  className?: string;
-  horizontal?: boolean;
-};
-
-export const Fieldset: FC<FieldsetProps> = ({
-  label,
-  htmlFor,
-  children,
-  className,
-  horizontal,
-}) => (
-  <fieldset
-    className={twMergeClsx(
-      "mb-2 flex gap-2",
-      className,
-      horizontal ? "flex-row-reverse items-center" : "flex-col"
-    )}
-  >
-    <label className="grow text-sm" htmlFor={htmlFor}>
-      {label}
-    </label>
-    {children}
-  </fieldset>
 );
 
 export const ModalFooter = () => {
